* Updating .jshintrc with `beforeEach` and `afterEach` values * Organizing the client and service files into a logical hierarchy * Fixing linting issues in a few files * Updating all references to the hierarchy change above
