+++ src/badges/client/index.js
