+++ test/app.js
@@ -3,8 +3,8 @@ var assert = require('assert');
 
 var testEnv = require('../src/environments');
 
-var badgeClient = require('../src/badgeClient.js')(testEnv);
-var badgeService = require('../src/badgeService.js');
+var badgeClient = require('../src/badges/client')(testEnv);
+var badgeService = require('../src/badges/service');
 badgeService.init(badgeClient, testEnv);
 
 var app = require('../src/app.js');
